♻️ refactor: refactor the auth action and common action
